feat: Update .gitignore to include .notes directory
docs: Revise AGENTS.md for universal guidelines on Kubex Agents

refactor: Simplify common.go by delegating banner logic to info package

fix: Update main.go to use logger for error handling

chore: Modify config.yml to use environment variable for GEMINI_API_KEY

feat: Implement GetDescriptions function in application.go for banner and description handling

docs: Enhance module.go descriptions for Analyzer functionality

chore: Add indexeddb-gemx-db-1758350108906 copy.json for project state management

chore: Add indexeddb-gemx-db-1758350108906.json for project state management

chore: Initialize package.json and package-lock.json for project dependencies
