+++ AGENTS.md
@@ -1,65 +1,55 @@
-# Repository Guidelines
+# Kubex AGENTS.md (Universal)
 
-## Project Structure & Module Organization
+Este documento serve como guia universal para configuração dos **Agents** em todos os repositórios Kubex. Ele unifica padrões já estabelecidos em cada projeto, evitando divergências e garantindo consistência para automações, bots e contribuidores humanos.
 
-- `cmd/cli`: CLI commands (Cobra).
-- `internal`: core packages (engine, types, module, providers).
-- `factory`: public-facing abstractions (providers, templates, engine shims).
-- `mocks`: centralized test doubles for `types.IAPIConfig` and `types.IConfig`.
-- `tests/tests_*`: external test packages (e.g., `tests_types`, `tests_engine`).
-- `docs`, `frontend`, `dist`, `support`: documentation, UI, build artifacts, scripts.
+## Contexto
 
-## Build, Test, and Development Commands
+Os agentes (Claude, Codex, Copilot, etc.) leem este documento como referência de **estrutura, padrões e práticas vigentes** no ecossistema Kubex. O objetivo é reduzir erros de interpretação e facilitar a manutenção de múltiplos repositórios.
 
-- `make build-dev linux amd64`: builds the binary to `dist/` using the install script.
-- `make test`: runs the Go test suite.
-- `make clean`: removes build artifacts.
-- `go test ./... -v`: run tests directly.
-- `go test ./... -cover`: run with coverage.
+## Project Standards (Current Reality)
 
-## Coding Style & Naming Conventions
+- **Manifest.json**: por padrão em `internal/module/info/manifest.json`. Se movido, é necessário atualizar manualmente as referências.
+- **Wrappers de módulo**: `internal/module/module.go` é a estrutura principal; `internal/module/wrpr.go` contém wrappers auxiliares.
+- **cmd/**: `main.go` atua como entrypoint principal. `cmd/cli/` guarda os entrypoints para wrappers de comandos CLI.
+- **Envvars**: sempre com fallback para valores padrão, garantindo resiliência.
+- **Carregamento de configs**: não ocorre no `main.go`; cada comando em `cmd/cli/` carrega apenas o necessário.
+- **READMEs**: todos os projetos têm `README.md` em inglês e opcional `docs/README.pt-BR.md`, com link no README em inglês.
+- **Makefile e support/**: genéricos, reagem às chaves definidas no `manifest.json`.
+- **Mocks**: sempre centralizados em módulos específicos, nunca hardcoded em lógicas reais.
+- **Logger universal (logz)**: todos os projetos usam o wrapper padrão em `internal/module/logger`. Recomenda-se importar com alias para consistência:
 
-- Language: Go 1.25; use `gofmt` (tabs, standard imports).
-- Packages: keep cohesive, use `internal/*` for non-public.
-- Test packages: place in `tests/tests_[package]`, name as `package <target>_test`.
-- Import aliases in tests: use stable, descriptive aliases (e.g., `t "github.com/kubex-ecosystem/analyzer/internal/types"`, `eng`, `prov`, `tpl`, `cli`, `mod`, `m` for `mocks`).
+  ```go
+  import (
+      gl "github.com/kubex-ecosystem/analyzer/internal/module/logger"
+  )
+  ```
 
-## Testing Guidelines
+- **Wrapper RegX**: todo módulo possui `internal/module/wrpr.go`, que contém exclusivamente o wrapper `func RegX() *[NOME_DO_MODULO]`. Isso padroniza o acesso ao módulo e mantém o `module.go` livre para customizações específicas sem risco de alterar o wrapper global.
 
-- Framework: Go `testing` (no third-party required).
-- Mocks: use centralized types in `mocks/` (avoid inline test doubles).
-- Coverage: aim for 70%+ overall; prefer unit tests that avoid network/IO.
-- Network tests: optional and gated by env (e.g., `GROMPT_E2E_GEMINI=1`, `GEMINI_API_KEY`).
-- Naming: use table-driven tests where helpful; file suffix `_test.go`.
+- **Internal como núcleo**: toda lógica central do projeto, ou qualquer parte que trate de algo sensível, crítico ou de complexidade média‑alta/alta, deve ficar em `internal/`.
+  - Se essa lógica precisar ser exposta para outro módulo ou uso externo, deve ser feita via **interfaces**, com construtores que permitam a instanciação dessas interfaces.
+  - A exportação deve ocorrer preferencialmente em `api/` ou `factory/`. Somente em casos específicos (como CLI ou domínios não ligados ao `internal/`) a exportação poderá estar fora desses packages.
 
-## Commit & Pull Request Guidelines
+- **Interface universal dos módulos**: o arquivo `internal/module/module.go` implementa a interface comum de **todos os módulos Kubex**. Cada módulo segue esse padrão, com métodos como `Alias()`, `ShortDescription()`, `LongDescription()`, `Usage()`, `Examples()`, `Active()`, `Module()`, `Execute()` e `Command()`. Essa estrutura garante consistência entre módulos e integração fluida com Cobra para CLI.
 
-- Commits: concise, imperative mood (e.g., "Add engine history tests").
-- Scope: prefer small, focused changes; include tests for behavior changes.
-- PRs: include description, motivation, and testing notes; link issues; add screenshots for CLI UX if relevant.
+  Exemplo simplificado:
 
-## Security & Configuration Tips
+  ```go
+  type Analyzer struct {
+      parentCmdName string
+      hideBanner   bool
+  }
 
-- Never commit API keys. Configure via env: `OPENAI_API_KEY`, `CLAUDE_API_KEY`, `GEMINI_API_KEY`, `DEEPSEEK_API_KEY`, `CHATGPT_API_KEY`, `OLLAMA_ENDPOINT`.
-- `.env` may aid local runs; keep it out of version control if sensitive.
+  func (m *Analyzer) Alias() string           { return "" }
+  func (m *Analyzer) ShortDescription() string { return "AI tools help in the editor, but they stop antes do PR, lacking governance." }
+  func (m *Analyzer) LongDescription() string  { return `Analyzer: An AI-powered tool...` }
+  func (m *Analyzer) Usage() string            { return "Analyzer [command] [args]" }
+  func (m *Analyzer) Active() bool             { return true }
+  func (m *Analyzer) Module() string           { return "Analyzer" }
+  func (m *Analyzer) Execute() error           { return m.Command().Execute() }
+  // ... demais métodos garantindo padronização
+  ```
 
-## Agent-Specific Instructions
+- **Banners**: todos os banners estão em `internal/module/info/application.go`, junto com o método auxiliar que cuida da lógica de impressão. Eles não ficam em `module/` porque também são usados pelos comandos em `cmd/cli/`. Se estivessem em `main.go`, criariam dependência cíclica, por isso foram isolados nesse arquivo específico.
 
-- When adding tests for new providers, reuse `mocks.APIConfigMock` and validate via `types.ProviderImpl` to exercise shared contracts.
-- Prefer adding reusable helpers to `mocks/` rather than duplicating setup logic across `tests/tests_*` packages.
-
-## Architecture Overview (Checkpoint)
-
-- Engine: orchestrates prompts, templates, and providers; exposes `ProcessPrompt`, `BatchProcess`, and history; selects providers from config.
-- Types: defines `IConfig`, `IAPIConfig`, and `ProviderImpl`; implements concrete API clients (OpenAI, Claude, Gemini, DeepSeek, Ollama, ChatGPT).
-- Providers: constructors wiring `ProviderImpl` with an `IAPIConfig`; `factory/providers` exposes interfaces and `Initialize` helpers.
-- Templates: `factory/templates` offers a simple `Manager` for variable substitution and future extensions.
-- CLI/Module: Cobra commands in `cmd/cli`; `internal/module` sets usage, logging, and aggregates subcommands.
-
-Architecture checkpoints
-
-- Data flow: Config -> Engine -> ProviderImpl(VAPI) -> API client -> Response -> History.
-- Injection: prefer `IConfig`/`IAPIConfig` at boundaries; keep network in API clients only.
-- Tests: mock at `IAPIConfig` layer; avoid real network unless explicitly gated via env.
-
-For a rendered diagram, see docs/development/architecture.md (Mermaid).
+- **CLI design customizado**: todo design customizado do CLI (cores, layout e estilos de exibição) está centralizado em `internal/module/usage.go`. Esse arquivo define a aparência dos comandos e mensagens, mantendo o padrão visual consistente entre os módulos.
