+++ cmd/main.go
@@ -1,31 +1,13 @@
 package main
 
 import (
-	"log"
+	gl "github.com/kubex-ecosystem/analyzer/internal/module/logger"
 
 	"github.com/kubex-ecosystem/analyzer/internal/module"
 )
 
-// func getEnv(k, d string) string {
-// 	if v := os.Getenv(k); v != "" {
-// 		return v
-// 	}
-// 	return d
-// }
-
-func init() {
-	// lê ou não lê o arquivo de configuração no init do main?
-	// talvez seja melhor deixar para cada comando ler o arquivo de configuração
-
-	// cfgPath := getEnv("PROVIDERS_CFG", "config/config.example.yml")
-	// _, err := registry.Load(cfgPath)
-	// if err != nil {
-	// 	log.Fatal(err)
-	// }
-}
-
 func main() {
 	if err := module.RegX().Command().Execute(); err != nil {
-		log.Fatal(err)
+		gl.Log("fatal", err.Error())
 	}
 }
