+++ config/config.yml
@@ -7,17 +7,11 @@ providers:
     key_env: GROQ_API_KEY
     default_model: llama-3.1-8b-instant
 
-  # gemini:
-  #   type: gemini
-  #   base_url: https://generativelanguage.googleapis.com
-  #   key_env: "AIzaSyDmybu7EzXMt20jLB8WVi09kTErJYR_HKI"
-  #   default_model: "gemini-2.5-flash"
-
   gemini:
     type: gemini
     base_url: https://generativelanguage.googleapis.com
-    key_env: "AIzaSyDmybu7EzXMt20jLB8WVi09kTErJYR_HKI"
-    default_model: "gemini-2.0-flash-exp"
+    key_env: GEMINI_API_KEY
+    default_model: gemini-2.0-pro
 
   openai:
     type: openai
